Release v6.2

See merge request mobile/android/android!2278

+++ app/src/main/java/mega/privacy/android/app/getLink/GetLinkViewModel.kt
@@ -140,7 +140,7 @@ class GetLinkViewModel @Inject constructor(
      *
      * @param expiryDate Expiry date to export.
      */
-    fun exportWithTimestamp(expiryDate: Int) {
+    fun exportWithTimestamp(expiryDate: Long) {
         exportNodeUseCase.export(node, expiryDate)
             .subscribeOn(Schedulers.io())
             .observeOn(AndroidSchedulers.mainThread())
